app/requirements.txt
fastapi[standard]
uvicorn==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pydantic[email]==2.10.3
itsdangerous==2.1.2
redis==5.2.0
authlib==1.3.2
httpx==0.28.1
celery==5.3.4
